# Conflicts:
#	CHANGELOG.md - Solved manually
#	Entities/AHuman.cpp - Used theirs, but kept my item in reach fix
#	Entities/Activity.cpp - Used theirs, but kept my playerNum fix
#	Entities/MovableObject.cpp - Solved manually, copied the script running changes from theirs
#	Managers/ActivityMan.cpp - Solved manually, mostly used mine
#	RTEA.vcxproj.filters - Solved manually

+++ Activities/GATutorial.h
@@ -120,7 +120,7 @@ ClassInfoGetters
 //                  but only for a limited number of teams. If -1, not applicable.
 // Return value:    Whether the Scene has the right stuff.
 
-	bool SceneIsCompatible(Scene *pScene, short teams = -1) override;
+	bool SceneIsCompatible(Scene *pScene, int teams = -1) override;
 
 
 //////////////////////////////////////////////////////////////////////////////////////////
